fix: BS소모품 발주수량 산출 수정

See merge request pjt/bsp/sms/sms-wells-back-app!2500

+++ src/main/java/com/kyowon/sms/wells/web/service/stock/dvo/WsnaBsCsmbGiveAOrderDvo.java
@@ -24,7 +24,7 @@ public class WsnaBsCsmbGiveAOrderDvo {
     int pajuLgstCnrStocQty;
     int sgsuLgstCnrStocQty;
     int woStocQty;
-    int stocPersMmN;
+    float stocPersMmN;
     String etExsDt;
     int goUprc;
     int ncstQty;
